fix(stability), #427: Set precision of error functions to float32 to correct the stability of the solution on different computers.

+++ co2mpas/model/physical/engine/co2_emission.py
@@ -1788,7 +1788,7 @@ def calibrate_model_params(
     min_e_and_p = [np.inf, copy.deepcopy(params)]
 
     def error_func(params, *args, **kwargs):
-        res = error_f(params, *args, **kwargs)
+        res = np.float32(error_f(params, *args, **kwargs))
 
         if res < min_e_and_p[0]:
             min_e_and_p[0], min_e_and_p[1] = (res, copy.deepcopy(params))
